backend/requirements.txt
flask-restful
langchain-ollama
bs4